Split tool handlers into separate files
