+++ codex-rs/core/src/tools/handlers/goal.rs
@@ -5,28 +5,20 @@
 //! the existing goal complete.
 
 use crate::function_tool::FunctionCallError;
-use crate::goals::CreateGoalRequest;
-use crate::goals::GoalRuntimeEvent;
-use crate::goals::SetGoalRequest;
 use crate::tools::context::FunctionToolOutput;
-use crate::tools::context::ToolInvocation;
-use crate::tools::context::ToolPayload;
-use crate::tools::handlers::parse_arguments;
-use crate::tools::registry::ToolHandler;
-use crate::tools::registry::ToolKind;
 use codex_protocol::protocol::ThreadGoal;
 use codex_protocol::protocol::ThreadGoalStatus;
-use codex_tools::CREATE_GOAL_TOOL_NAME;
-use codex_tools::GET_GOAL_TOOL_NAME;
-use codex_tools::ToolName;
-use codex_tools::UPDATE_GOAL_TOOL_NAME;
 use serde::Deserialize;
 use serde::Serialize;
 use std::fmt::Write as _;
 
-pub struct GetGoalHandler;
-pub struct CreateGoalHandler;
-pub struct UpdateGoalHandler;
+mod create_goal;
+mod get_goal;
+mod update_goal;
+
+pub use create_goal::CreateGoalHandler;
+pub use get_goal::GetGoalHandler;
+pub use update_goal::UpdateGoalHandler;
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "snake_case")]
@@ -73,148 +65,6 @@ impl GoalToolResponse {
             remaining_tokens,
             completion_budget_report,
         }
-    }
-}
-
-impl ToolHandler for GetGoalHandler {
-    type Output = FunctionToolOutput;
-
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain(GET_GOAL_TOOL_NAME)
-    }
-
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session, payload, ..
-        } = invocation;
-
-        match payload {
-            ToolPayload::Function { .. } => {
-                let goal = session
-                    .get_thread_goal()
-                    .await
-                    .map_err(|err| FunctionCallError::RespondToModel(format_goal_error(err)))?;
-                goal_response(goal, CompletionBudgetReport::Omit)
-            }
-            _ => Err(FunctionCallError::RespondToModel(
-                "get_goal handler received unsupported payload".to_string(),
-            )),
-        }
-    }
-}
-
-impl ToolHandler for CreateGoalHandler {
-    type Output = FunctionToolOutput;
-
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain(CREATE_GOAL_TOOL_NAME)
-    }
-
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session,
-            turn,
-            payload,
-            ..
-        } = invocation;
-
-        let arguments = match payload {
-            ToolPayload::Function { arguments } => arguments,
-            _ => {
-                return Err(FunctionCallError::RespondToModel(
-                    "goal handler received unsupported payload".to_string(),
-                ));
-            }
-        };
-
-        let args: CreateGoalArgs = parse_arguments(&arguments)?;
-        let goal = session
-            .create_thread_goal(
-                turn.as_ref(),
-                CreateGoalRequest {
-                    objective: args.objective,
-                    token_budget: args.token_budget,
-                },
-            )
-            .await
-            .map_err(|err| {
-                if err
-                    .chain()
-                    .any(|cause| cause.to_string().contains("already has a goal"))
-                {
-                    FunctionCallError::RespondToModel(
-                        "cannot create a new goal because this thread already has a goal; use update_goal only when the existing goal is complete"
-                            .to_string(),
-                    )
-                } else {
-                    FunctionCallError::RespondToModel(format_goal_error(err))
-                }
-            })?;
-        goal_response(Some(goal), CompletionBudgetReport::Omit)
-    }
-}
-
-impl ToolHandler for UpdateGoalHandler {
-    type Output = FunctionToolOutput;
-
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain(UPDATE_GOAL_TOOL_NAME)
-    }
-
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session,
-            turn,
-            payload,
-            ..
-        } = invocation;
-
-        let arguments = match payload {
-            ToolPayload::Function { arguments } => arguments,
-            _ => {
-                return Err(FunctionCallError::RespondToModel(
-                    "update_goal handler received unsupported payload".to_string(),
-                ));
-            }
-        };
-
-        let args: UpdateGoalArgs = parse_arguments(&arguments)?;
-        if args.status != ThreadGoalStatus::Complete {
-            return Err(FunctionCallError::RespondToModel(
-                "update_goal can only mark the existing goal complete; pause, resume, and budget-limited status changes are controlled by the user or system"
-                    .to_string(),
-            ));
-        }
-        session
-            .goal_runtime_apply(GoalRuntimeEvent::ToolCompletedGoal {
-                turn_context: turn.as_ref(),
-            })
-            .await
-            .map_err(|err| FunctionCallError::RespondToModel(format_goal_error(err)))?;
-        let goal = session
-            .set_thread_goal(
-                turn.as_ref(),
-                SetGoalRequest {
-                    objective: None,
-                    status: Some(ThreadGoalStatus::Complete),
-                    token_budget: None,
-                },
-            )
-            .await
-            .map_err(|err| FunctionCallError::RespondToModel(format_goal_error(err)))?;
-        goal_response(Some(goal), CompletionBudgetReport::Include)
     }
 }
