+++ codex-rs/core/src/tools/handlers/shell.rs
@@ -1,17 +1,13 @@
-use codex_protocol::ThreadId;
+use codex_features::Feature;
 use codex_protocol::models::ShellCommandToolCallParams;
 use codex_protocol::models::ShellToolCallParams;
 use serde_json::Value as JsonValue;
 use std::sync::Arc;
 
-use crate::exec::ExecCapturePolicy;
 use crate::exec::ExecParams;
-use crate::exec_env::create_env;
 use crate::exec_policy::ExecApprovalRequest;
 use crate::function_tool::FunctionCallError;
-use crate::maybe_emit_implicit_skill_invocation;
 use crate::session::turn_context::TurnContext;
-use crate::shell::Shell;
 use crate::tools::context::FunctionToolOutput;
 use crate::tools::context::ToolInvocation;
 use crate::tools::context::ToolOutput;
@@ -23,38 +19,26 @@ use crate::tools::handlers::apply_patch::intercept_apply_patch;
 use crate::tools::handlers::implicit_granted_permissions;
 use crate::tools::handlers::normalize_and_validate_additional_permissions;
 use crate::tools::handlers::parse_arguments;
-use crate::tools::handlers::parse_arguments_with_base_path;
-use crate::tools::handlers::resolve_workdir_base_path;
 use crate::tools::hook_names::HookToolName;
 use crate::tools::orchestrator::ToolOrchestrator;
 use crate::tools::registry::PostToolUsePayload;
 use crate::tools::registry::PreToolUsePayload;
-use crate::tools::registry::ToolHandler;
-use crate::tools::registry::ToolKind;
 use crate::tools::runtimes::shell::ShellRequest;
 use crate::tools::runtimes::shell::ShellRuntime;
 use crate::tools::runtimes::shell::ShellRuntimeBackend;
 use crate::tools::sandboxing::ToolCtx;
-use codex_features::Feature;
 use codex_protocol::models::AdditionalPermissionProfile;
 use codex_protocol::protocol::ExecCommandSource;
-use codex_shell_command::is_safe_command::is_known_safe_command;
-use codex_tools::ShellCommandBackendConfig;
-use codex_tools::ToolName;
 
-pub struct ShellHandler;
-pub struct ContainerExecHandler;
-pub struct LocalShellHandler;
+mod container_exec;
+mod local_shell;
+mod shell_command;
+mod shell_handler;
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum ShellCommandBackend {
-    Classic,
-    ZshFork,
-}
-
-pub struct ShellCommandHandler {
-    backend: ShellCommandBackend,
-}
+pub use container_exec::ContainerExecHandler;
+pub use local_shell::LocalShellHandler;
+pub use shell_command::ShellCommandHandler;
+pub use shell_handler::ShellHandler;
 
 fn shell_function_payload_command(payload: &ToolPayload) -> Option<String> {
     let ToolPayload::Function { arguments } = payload else {
@@ -100,333 +84,6 @@ struct RunExecLikeArgs {
     shell_runtime_backend: ShellRuntimeBackend,
 }
 
-impl ShellHandler {
-    fn to_exec_params(
-        params: &ShellToolCallParams,
-        turn_context: &TurnContext,
-        thread_id: ThreadId,
-    ) -> ExecParams {
-        ExecParams {
-            command: params.command.clone(),
-            cwd: turn_context.resolve_path(params.workdir.clone()),
-            expiration: params.timeout_ms.into(),
-            capture_policy: ExecCapturePolicy::ShellTool,
-            env: create_env(&turn_context.shell_environment_policy, Some(thread_id)),
-            network: turn_context.network.clone(),
-            sandbox_permissions: params.sandbox_permissions.unwrap_or_default(),
-            windows_sandbox_level: turn_context.windows_sandbox_level,
-            windows_sandbox_private_desktop: turn_context
-                .config
-                .permissions
-                .windows_sandbox_private_desktop,
-            justification: params.justification.clone(),
-            arg0: None,
-        }
-    }
-}
-
-impl ShellCommandHandler {
-    fn shell_runtime_backend(&self) -> ShellRuntimeBackend {
-        match self.backend {
-            ShellCommandBackend::Classic => ShellRuntimeBackend::ShellCommandClassic,
-            ShellCommandBackend::ZshFork => ShellRuntimeBackend::ShellCommandZshFork,
-        }
-    }
-
-    fn resolve_use_login_shell(
-        login: Option<bool>,
-        allow_login_shell: bool,
-    ) -> Result<bool, FunctionCallError> {
-        if !allow_login_shell && login == Some(true) {
-            return Err(FunctionCallError::RespondToModel(
-                "login shell is disabled by config; omit `login` or set it to false.".to_string(),
-            ));
-        }
-
-        Ok(login.unwrap_or(allow_login_shell))
-    }
-
-    fn base_command(shell: &Shell, command: &str, use_login_shell: bool) -> Vec<String> {
-        shell.derive_exec_args(command, use_login_shell)
-    }
-
-    fn to_exec_params(
-        params: &ShellCommandToolCallParams,
-        session: &crate::session::session::Session,
-        turn_context: &TurnContext,
-        thread_id: ThreadId,
-        allow_login_shell: bool,
-    ) -> Result<ExecParams, FunctionCallError> {
-        let shell = session.user_shell();
-        let use_login_shell = Self::resolve_use_login_shell(params.login, allow_login_shell)?;
-        let command = Self::base_command(shell.as_ref(), &params.command, use_login_shell);
-
-        Ok(ExecParams {
-            command,
-            cwd: turn_context.resolve_path(params.workdir.clone()),
-            expiration: params.timeout_ms.into(),
-            capture_policy: ExecCapturePolicy::ShellTool,
-            env: create_env(&turn_context.shell_environment_policy, Some(thread_id)),
-            network: turn_context.network.clone(),
-            sandbox_permissions: params.sandbox_permissions.unwrap_or_default(),
-            windows_sandbox_level: turn_context.windows_sandbox_level,
-            windows_sandbox_private_desktop: turn_context
-                .config
-                .permissions
-                .windows_sandbox_private_desktop,
-            justification: params.justification.clone(),
-            arg0: None,
-        })
-    }
-}
-
-impl From<ShellCommandBackendConfig> for ShellCommandHandler {
-    fn from(config: ShellCommandBackendConfig) -> Self {
-        let backend = match config {
-            ShellCommandBackendConfig::Classic => ShellCommandBackend::Classic,
-            ShellCommandBackendConfig::ZshFork => ShellCommandBackend::ZshFork,
-        };
-        Self { backend }
-    }
-}
-
-impl ToolHandler for ShellHandler {
-    type Output = FunctionToolOutput;
-
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain("shell")
-    }
-
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    fn matches_kind(&self, payload: &ToolPayload) -> bool {
-        matches!(payload, ToolPayload::Function { .. })
-    }
-
-    async fn is_mutating(&self, invocation: &ToolInvocation) -> bool {
-        let ToolPayload::Function { arguments } = &invocation.payload else {
-            return true;
-        };
-
-        serde_json::from_str::<ShellToolCallParams>(arguments)
-            .map(|params| !is_known_safe_command(&params.command))
-            .unwrap_or(true)
-    }
-
-    fn pre_tool_use_payload(&self, invocation: &ToolInvocation) -> Option<PreToolUsePayload> {
-        shell_function_pre_tool_use_payload(invocation)
-    }
-
-    fn post_tool_use_payload(
-        &self,
-        invocation: &ToolInvocation,
-        result: &Self::Output,
-    ) -> Option<PostToolUsePayload> {
-        shell_function_post_tool_use_payload(invocation, result)
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session,
-            turn,
-            tracker,
-            call_id,
-            payload,
-            ..
-        } = invocation;
-
-        let arguments = match payload {
-            ToolPayload::Function { arguments } => arguments,
-            _ => {
-                return Err(FunctionCallError::RespondToModel(
-                    "unsupported payload for shell handler".to_string(),
-                ));
-            }
-        };
-
-        let cwd = resolve_workdir_base_path(&arguments, &turn.cwd)?;
-        let params: ShellToolCallParams = parse_arguments_with_base_path(&arguments, &cwd)?;
-        let prefix_rule = params.prefix_rule.clone();
-        let exec_params =
-            ShellHandler::to_exec_params(&params, turn.as_ref(), session.conversation_id);
-        ShellHandler::run_exec_like(RunExecLikeArgs {
-            tool_name: "shell".to_string(),
-            exec_params,
-            hook_command: codex_shell_command::parse_command::shlex_join(&params.command),
-            additional_permissions: params.additional_permissions.clone(),
-            prefix_rule,
-            session,
-            turn,
-            tracker,
-            call_id,
-            freeform: false,
-            shell_runtime_backend: ShellRuntimeBackend::Generic,
-        })
-        .await
-    }
-}
-
-impl ToolHandler for ContainerExecHandler {
-    type Output = FunctionToolOutput;
-
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain("container.exec")
-    }
-
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    fn matches_kind(&self, payload: &ToolPayload) -> bool {
-        matches!(payload, ToolPayload::Function { .. })
-    }
-
-    async fn is_mutating(&self, invocation: &ToolInvocation) -> bool {
-        let ToolPayload::Function { arguments } = &invocation.payload else {
-            return true;
-        };
-
-        serde_json::from_str::<ShellToolCallParams>(arguments)
-            .map(|params| !is_known_safe_command(&params.command))
-            .unwrap_or(true)
-    }
-
-    fn pre_tool_use_payload(&self, invocation: &ToolInvocation) -> Option<PreToolUsePayload> {
-        shell_function_pre_tool_use_payload(invocation)
-    }
-
-    fn post_tool_use_payload(
-        &self,
-        invocation: &ToolInvocation,
-        result: &Self::Output,
-    ) -> Option<PostToolUsePayload> {
-        shell_function_post_tool_use_payload(invocation, result)
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session,
-            turn,
-            tracker,
-            call_id,
-            payload,
-            ..
-        } = invocation;
-
-        let arguments = match payload {
-            ToolPayload::Function { arguments } => arguments,
-            _ => {
-                return Err(FunctionCallError::RespondToModel(
-                    "unsupported payload for container.exec handler".to_string(),
-                ));
-            }
-        };
-
-        let cwd = resolve_workdir_base_path(&arguments, &turn.cwd)?;
-        let params: ShellToolCallParams = parse_arguments_with_base_path(&arguments, &cwd)?;
-        let prefix_rule = params.prefix_rule.clone();
-        let exec_params =
-            ShellHandler::to_exec_params(&params, turn.as_ref(), session.conversation_id);
-        ShellHandler::run_exec_like(RunExecLikeArgs {
-            tool_name: "container.exec".to_string(),
-            exec_params,
-            hook_command: codex_shell_command::parse_command::shlex_join(&params.command),
-            additional_permissions: params.additional_permissions.clone(),
-            prefix_rule,
-            session,
-            turn,
-            tracker,
-            call_id,
-            freeform: false,
-            shell_runtime_backend: ShellRuntimeBackend::Generic,
-        })
-        .await
-    }
-}
-
-impl ToolHandler for LocalShellHandler {
-    type Output = FunctionToolOutput;
-
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain("local_shell")
-    }
-
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    fn matches_kind(&self, payload: &ToolPayload) -> bool {
-        matches!(payload, ToolPayload::LocalShell { .. })
-    }
-
-    async fn is_mutating(&self, invocation: &ToolInvocation) -> bool {
-        let ToolPayload::LocalShell { params } = &invocation.payload else {
-            return true;
-        };
-
-        !is_known_safe_command(&params.command)
-    }
-
-    fn pre_tool_use_payload(&self, invocation: &ToolInvocation) -> Option<PreToolUsePayload> {
-        local_shell_payload_command(&invocation.payload).map(|command| PreToolUsePayload {
-            tool_name: HookToolName::bash(),
-            tool_input: serde_json::json!({ "command": command }),
-        })
-    }
-
-    fn post_tool_use_payload(
-        &self,
-        invocation: &ToolInvocation,
-        result: &Self::Output,
-    ) -> Option<PostToolUsePayload> {
-        let tool_response =
-            result.post_tool_use_response(&invocation.call_id, &invocation.payload)?;
-        let command = local_shell_payload_command(&invocation.payload)?;
-        Some(PostToolUsePayload {
-            tool_name: HookToolName::bash(),
-            tool_use_id: invocation.call_id.clone(),
-            tool_input: serde_json::json!({ "command": command }),
-            tool_response,
-        })
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session,
-            turn,
-            tracker,
-            call_id,
-            payload,
-            ..
-        } = invocation;
-
-        let ToolPayload::LocalShell { params } = payload else {
-            return Err(FunctionCallError::RespondToModel(
-                "unsupported payload for local_shell handler".to_string(),
-            ));
-        };
-
-        let exec_params =
-            ShellHandler::to_exec_params(&params, turn.as_ref(), session.conversation_id);
-        ShellHandler::run_exec_like(RunExecLikeArgs {
-            tool_name: "local_shell".to_string(),
-            exec_params,
-            hook_command: codex_shell_command::parse_command::shlex_join(&params.command),
-            additional_permissions: None,
-            prefix_rule: None,
-            session,
-            turn,
-            tracker,
-            call_id,
-            freeform: false,
-            shell_runtime_backend: ShellRuntimeBackend::Generic,
-        })
-        .await
-    }
-}
-
 fn shell_function_pre_tool_use_payload(invocation: &ToolInvocation) -> Option<PreToolUsePayload> {
     shell_function_payload_command(&invocation.payload).map(|command| PreToolUsePayload {
         tool_name: HookToolName::bash(),
@@ -448,316 +105,203 @@ fn shell_function_post_tool_use_payload(
     })
 }
 
-impl ToolHandler for ShellCommandHandler {
-    type Output = FunctionToolOutput;
+async fn run_exec_like(args: RunExecLikeArgs) -> Result<FunctionToolOutput, FunctionCallError> {
+    let RunExecLikeArgs {
+        tool_name,
+        exec_params,
+        hook_command,
+        additional_permissions,
+        prefix_rule,
+        session,
+        turn,
+        tracker,
+        call_id,
+        freeform,
+        shell_runtime_backend,
+    } = args;
 
-    fn tool_name(&self) -> ToolName {
-        ToolName::plain("shell_command")
+    let mut exec_params = exec_params;
+    let Some(turn_environment) = turn.environments.primary() else {
+        return Err(FunctionCallError::RespondToModel(
+            "shell is unavailable in this session".to_string(),
+        ));
+    };
+    let fs = turn_environment.environment.get_filesystem();
+
+    let dependency_env = session.dependency_env().await;
+    if !dependency_env.is_empty() {
+        exec_params.env.extend(dependency_env.clone());
     }
 
-    fn kind(&self) -> ToolKind {
-        ToolKind::Function
-    }
-
-    fn matches_kind(&self, payload: &ToolPayload) -> bool {
-        matches!(payload, ToolPayload::Function { .. })
-    }
-
-    async fn is_mutating(&self, invocation: &ToolInvocation) -> bool {
-        let ToolPayload::Function { arguments } = &invocation.payload else {
-            return true;
-        };
-
-        serde_json::from_str::<ShellCommandToolCallParams>(arguments)
-            .map(|params| {
-                let use_login_shell = match Self::resolve_use_login_shell(
-                    params.login,
-                    invocation.turn.tools_config.allow_login_shell,
-                ) {
-                    Ok(use_login_shell) => use_login_shell,
-                    Err(_) => return true,
-                };
-                let shell = invocation.session.user_shell();
-                let command = Self::base_command(shell.as_ref(), &params.command, use_login_shell);
-                !is_known_safe_command(&command)
-            })
-            .unwrap_or(true)
-    }
-
-    fn pre_tool_use_payload(&self, invocation: &ToolInvocation) -> Option<PreToolUsePayload> {
-        shell_command_payload_command(&invocation.payload).map(|command| PreToolUsePayload {
-            tool_name: HookToolName::bash(),
-            tool_input: serde_json::json!({ "command": command }),
-        })
-    }
-
-    fn post_tool_use_payload(
-        &self,
-        invocation: &ToolInvocation,
-        result: &Self::Output,
-    ) -> Option<PostToolUsePayload> {
-        let tool_response =
-            result.post_tool_use_response(&invocation.call_id, &invocation.payload)?;
-        let command = shell_command_payload_command(&invocation.payload)?;
-        Some(PostToolUsePayload {
-            tool_name: HookToolName::bash(),
-            tool_use_id: invocation.call_id.clone(),
-            tool_input: serde_json::json!({ "command": command }),
-            tool_response,
-        })
-    }
-
-    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
-        let ToolInvocation {
-            session,
-            turn,
-            tracker,
-            call_id,
-            payload,
-            ..
-        } = invocation;
-
-        let ToolPayload::Function { arguments } = payload else {
-            return Err(FunctionCallError::RespondToModel(format!(
-                "unsupported payload for shell_command handler: {}",
-                self.tool_name().display()
-            )));
-        };
-
-        let cwd = resolve_workdir_base_path(&arguments, &turn.cwd)?;
-        let params: ShellCommandToolCallParams = parse_arguments_with_base_path(&arguments, &cwd)?;
-        let workdir = turn.resolve_path(params.workdir.clone());
-        maybe_emit_implicit_skill_invocation(
-            session.as_ref(),
-            turn.as_ref(),
-            &params.command,
-            &workdir,
-        )
-        .await;
-        let prefix_rule = params.prefix_rule.clone();
-        let exec_params = Self::to_exec_params(
-            &params,
-            session.as_ref(),
-            turn.as_ref(),
-            session.conversation_id,
-            turn.tools_config.allow_login_shell,
-        )?;
-        ShellHandler::run_exec_like(RunExecLikeArgs {
-            tool_name: self.tool_name().display(),
-            exec_params,
-            hook_command: params.command,
-            additional_permissions: params.additional_permissions.clone(),
-            prefix_rule,
-            session,
-            turn,
-            tracker,
-            call_id,
-            freeform: true,
-            shell_runtime_backend: self.shell_runtime_backend(),
-        })
-        .await
-    }
-}
-
-impl ShellHandler {
-    async fn run_exec_like(args: RunExecLikeArgs) -> Result<FunctionToolOutput, FunctionCallError> {
-        let RunExecLikeArgs {
-            tool_name,
-            exec_params,
-            hook_command,
-            additional_permissions,
-            prefix_rule,
-            session,
-            turn,
-            tracker,
-            call_id,
-            freeform,
-            shell_runtime_backend,
-        } = args;
-
-        let mut exec_params = exec_params;
-        let Some(turn_environment) = turn.environments.primary() else {
-            return Err(FunctionCallError::RespondToModel(
-                "shell is unavailable in this session".to_string(),
-            ));
-        };
-        let fs = turn_environment.environment.get_filesystem();
-
-        let dependency_env = session.dependency_env().await;
-        if !dependency_env.is_empty() {
-            exec_params.env.extend(dependency_env.clone());
+    let mut explicit_env_overrides = turn.shell_environment_policy.r#set.clone();
+    for key in dependency_env.keys() {
+        if let Some(value) = exec_params.env.get(key) {
+            explicit_env_overrides.insert(key.clone(), value.clone());
         }
+    }
 
-        let mut explicit_env_overrides = turn.shell_environment_policy.r#set.clone();
-        for key in dependency_env.keys() {
-            if let Some(value) = exec_params.env.get(key) {
-                explicit_env_overrides.insert(key.clone(), value.clone());
-            }
-        }
+    let exec_permission_approvals_enabled =
+        session.features().enabled(Feature::ExecPermissionApprovals);
+    let requested_additional_permissions = additional_permissions.clone();
+    let effective_additional_permissions = apply_granted_turn_permissions(
+        session.as_ref(),
+        turn.cwd.as_path(),
+        exec_params.sandbox_permissions,
+        additional_permissions,
+    )
+    .await;
+    let additional_permissions_allowed = exec_permission_approvals_enabled
+        || (session.features().enabled(Feature::RequestPermissionsTool)
+            && effective_additional_permissions.permissions_preapproved);
+    let normalized_additional_permissions = implicit_granted_permissions(
+        exec_params.sandbox_permissions,
+        requested_additional_permissions.as_ref(),
+        &effective_additional_permissions,
+    )
+    .map_or_else(
+        || {
+            normalize_and_validate_additional_permissions(
+                additional_permissions_allowed,
+                turn.approval_policy.value(),
+                effective_additional_permissions.sandbox_permissions,
+                effective_additional_permissions.additional_permissions,
+                effective_additional_permissions.permissions_preapproved,
+                &exec_params.cwd,
+            )
+        },
+        |permissions| Ok(Some(permissions)),
+    )
+    .map_err(FunctionCallError::RespondToModel)?;
 
-        let exec_permission_approvals_enabled =
-            session.features().enabled(Feature::ExecPermissionApprovals);
-        let requested_additional_permissions = additional_permissions.clone();
-        let effective_additional_permissions = apply_granted_turn_permissions(
-            session.as_ref(),
-            turn.cwd.as_path(),
-            exec_params.sandbox_permissions,
-            additional_permissions,
+    // Approval policy guard for explicit escalation in non-OnRequest modes.
+    // Sticky turn permissions have already been approved, so they should
+    // continue through the normal exec approval flow for the command.
+    if effective_additional_permissions
+        .sandbox_permissions
+        .requests_sandbox_override()
+        && !effective_additional_permissions.permissions_preapproved
+        && !matches!(
+            turn.approval_policy.value(),
+            codex_protocol::protocol::AskForApproval::OnRequest
         )
-        .await;
-        let additional_permissions_allowed = exec_permission_approvals_enabled
-            || (session.features().enabled(Feature::RequestPermissionsTool)
-                && effective_additional_permissions.permissions_preapproved);
-        let normalized_additional_permissions = implicit_granted_permissions(
-            exec_params.sandbox_permissions,
-            requested_additional_permissions.as_ref(),
-            &effective_additional_permissions,
-        )
-        .map_or_else(
-            || {
-                normalize_and_validate_additional_permissions(
-                    additional_permissions_allowed,
-                    turn.approval_policy.value(),
-                    effective_additional_permissions.sandbox_permissions,
-                    effective_additional_permissions.additional_permissions,
-                    effective_additional_permissions.permissions_preapproved,
-                    &exec_params.cwd,
-                )
+    {
+        let approval_policy = turn.approval_policy.value();
+        return Err(FunctionCallError::RespondToModel(format!(
+            "approval policy is {approval_policy:?}; reject command — you should not ask for escalated permissions if the approval policy is {approval_policy:?}"
+        )));
+    }
+
+    // Intercept apply_patch if present.
+    if let Some(output) = intercept_apply_patch(
+        &exec_params.command,
+        &exec_params.cwd,
+        fs.as_ref(),
+        session.clone(),
+        turn.clone(),
+        Some(&tracker),
+        &call_id,
+        tool_name.as_str(),
+    )
+    .await?
+    {
+        return Ok(output);
+    }
+
+    let source = ExecCommandSource::Agent;
+    let emitter = ToolEmitter::shell(
+        exec_params.command.clone(),
+        exec_params.cwd.clone(),
+        source,
+        freeform,
+    );
+    let event_ctx = ToolEventCtx::new(
+        session.as_ref(),
+        turn.as_ref(),
+        &call_id,
+        /*turn_diff_tracker*/ None,
+    );
+    emitter.begin(event_ctx).await;
+
+    let file_system_sandbox_policy = turn.file_system_sandbox_policy();
+    let exec_approval_requirement = session
+        .services
+        .exec_policy
+        .create_exec_approval_requirement_for_command(ExecApprovalRequest {
+            command: &exec_params.command,
+            approval_policy: turn.approval_policy.value(),
+            permission_profile: turn.permission_profile(),
+            file_system_sandbox_policy: &file_system_sandbox_policy,
+            sandbox_cwd: turn.cwd.as_path(),
+            sandbox_permissions: if effective_additional_permissions.permissions_preapproved {
+                codex_protocol::models::SandboxPermissions::UseDefault
+            } else {
+                effective_additional_permissions.sandbox_permissions
             },
-            |permissions| Ok(Some(permissions)),
-        )
-        .map_err(FunctionCallError::RespondToModel)?;
-
-        // Approval policy guard for explicit escalation in non-OnRequest modes.
-        // Sticky turn permissions have already been approved, so they should
-        // continue through the normal exec approval flow for the command.
-        if effective_additional_permissions
-            .sandbox_permissions
-            .requests_sandbox_override()
-            && !effective_additional_permissions.permissions_preapproved
-            && !matches!(
-                turn.approval_policy.value(),
-                codex_protocol::protocol::AskForApproval::OnRequest
-            )
-        {
-            let approval_policy = turn.approval_policy.value();
-            return Err(FunctionCallError::RespondToModel(format!(
-                "approval policy is {approval_policy:?}; reject command — you should not ask for escalated permissions if the approval policy is {approval_policy:?}"
-            )));
-        }
-
-        // Intercept apply_patch if present.
-        if let Some(output) = intercept_apply_patch(
-            &exec_params.command,
-            &exec_params.cwd,
-            fs.as_ref(),
-            session.clone(),
-            turn.clone(),
-            Some(&tracker),
-            &call_id,
-            tool_name.as_str(),
-        )
-        .await?
-        {
-            return Ok(output);
-        }
-
-        let source = ExecCommandSource::Agent;
-        let emitter = ToolEmitter::shell(
-            exec_params.command.clone(),
-            exec_params.cwd.clone(),
-            source,
-            freeform,
-        );
-        let event_ctx = ToolEventCtx::new(
-            session.as_ref(),
-            turn.as_ref(),
-            &call_id,
-            /*turn_diff_tracker*/ None,
-        );
-        emitter.begin(event_ctx).await;
-
-        let file_system_sandbox_policy = turn.file_system_sandbox_policy();
-        let exec_approval_requirement = session
-            .services
-            .exec_policy
-            .create_exec_approval_requirement_for_command(ExecApprovalRequest {
-                command: &exec_params.command,
-                approval_policy: turn.approval_policy.value(),
-                permission_profile: turn.permission_profile(),
-                file_system_sandbox_policy: &file_system_sandbox_policy,
-                sandbox_cwd: turn.cwd.as_path(),
-                sandbox_permissions: if effective_additional_permissions.permissions_preapproved {
-                    codex_protocol::models::SandboxPermissions::UseDefault
-                } else {
-                    effective_additional_permissions.sandbox_permissions
-                },
-                prefix_rule,
-            })
-            .await;
-
-        let req = ShellRequest {
-            command: exec_params.command.clone(),
-            hook_command,
-            cwd: exec_params.cwd.clone(),
-            timeout_ms: exec_params.expiration.timeout_ms(),
-            env: exec_params.env.clone(),
-            explicit_env_overrides,
-            network: exec_params.network.clone(),
-            sandbox_permissions: effective_additional_permissions.sandbox_permissions,
-            additional_permissions: normalized_additional_permissions,
-            #[cfg(unix)]
-            additional_permissions_preapproved: effective_additional_permissions
-                .permissions_preapproved,
-            justification: exec_params.justification.clone(),
-            exec_approval_requirement,
-        };
-        let mut orchestrator = ToolOrchestrator::new();
-        let mut runtime = {
-            use ShellRuntimeBackend::*;
-            match shell_runtime_backend {
-                Generic => ShellRuntime::new(),
-                backend @ (ShellCommandClassic | ShellCommandZshFork) => {
-                    ShellRuntime::for_shell_command(backend)
-                }
-            }
-        };
-        let tool_ctx = ToolCtx {
-            session: session.clone(),
-            turn: turn.clone(),
-            call_id: call_id.clone(),
-            tool_name,
-        };
-        let out = orchestrator
-            .run(
-                &mut runtime,
-                &req,
-                &tool_ctx,
-                &turn,
-                turn.approval_policy.value(),
-            )
-            .await
-            .map(|result| result.output);
-        let event_ctx = ToolEventCtx::new(
-            session.as_ref(),
-            turn.as_ref(),
-            &call_id,
-            /*turn_diff_tracker*/ None,
-        );
-        let post_tool_use_response = out
-            .as_ref()
-            .ok()
-            .map(|output| crate::tools::format_exec_output_str(output, turn.truncation_policy))
-            .map(JsonValue::String);
-        let content = emitter.finish(event_ctx, out).await?;
-        Ok(FunctionToolOutput {
-            body: vec![
-                codex_protocol::models::FunctionCallOutputContentItem::InputText { text: content },
-            ],
-            success: Some(true),
-            post_tool_use_response,
+            prefix_rule,
         })
-    }
+        .await;
+
+    let req = ShellRequest {
+        command: exec_params.command.clone(),
+        hook_command,
+        cwd: exec_params.cwd.clone(),
+        timeout_ms: exec_params.expiration.timeout_ms(),
+        env: exec_params.env.clone(),
+        explicit_env_overrides,
+        network: exec_params.network.clone(),
+        sandbox_permissions: effective_additional_permissions.sandbox_permissions,
+        additional_permissions: normalized_additional_permissions,
+        #[cfg(unix)]
+        additional_permissions_preapproved: effective_additional_permissions
+            .permissions_preapproved,
+        justification: exec_params.justification.clone(),
+        exec_approval_requirement,
+    };
+    let mut orchestrator = ToolOrchestrator::new();
+    let mut runtime = {
+        use ShellRuntimeBackend::*;
+        match shell_runtime_backend {
+            Generic => ShellRuntime::new(),
+            backend @ (ShellCommandClassic | ShellCommandZshFork) => {
+                ShellRuntime::for_shell_command(backend)
+            }
+        }
+    };
+    let tool_ctx = ToolCtx {
+        session: session.clone(),
+        turn: turn.clone(),
+        call_id: call_id.clone(),
+        tool_name,
+    };
+    let out = orchestrator
+        .run(
+            &mut runtime,
+            &req,
+            &tool_ctx,
+            &turn,
+            turn.approval_policy.value(),
+        )
+        .await
+        .map(|result| result.output);
+    let event_ctx = ToolEventCtx::new(
+        session.as_ref(),
+        turn.as_ref(),
+        &call_id,
+        /*turn_diff_tracker*/ None,
+    );
+    let post_tool_use_response = out
+        .as_ref()
+        .ok()
+        .map(|output| crate::tools::format_exec_output_str(output, turn.truncation_policy))
+        .map(JsonValue::String);
+    let content = emitter.finish(event_ctx, out).await?;
+    Ok(FunctionToolOutput {
+        body: vec![
+            codex_protocol::models::FunctionCallOutputContentItem::InputText { text: content },
+        ],
+        success: Some(true),
+        post_tool_use_response,
+    })
 }
 
 #[cfg(test)]
