+++ codex-rs/core/src/tools/handlers/unified_exec/exec_command.rs
@@ -1,0 +1,309 @@
+use std::sync::Arc;
+
+use crate::function_tool::FunctionCallError;
+use crate::maybe_emit_implicit_skill_invocation;
+use crate::tools::context::ExecCommandToolOutput;
+use crate::tools::context::ToolInvocation;
+use crate::tools::context::ToolPayload;
+use crate::tools::handlers::apply_granted_turn_permissions;
+use crate::tools::handlers::apply_patch::intercept_apply_patch;
+use crate::tools::handlers::implicit_granted_permissions;
+use crate::tools::handlers::normalize_and_validate_additional_permissions;
+use crate::tools::handlers::parse_arguments;
+use crate::tools::handlers::parse_arguments_with_base_path;
+use crate::tools::handlers::resolve_tool_environment;
+use crate::tools::hook_names::HookToolName;
+use crate::tools::registry::PostToolUsePayload;
+use crate::tools::registry::PreToolUsePayload;
+use crate::tools::registry::ToolHandler;
+use crate::tools::registry::ToolKind;
+use crate::unified_exec::ExecCommandRequest;
+use crate::unified_exec::UnifiedExecContext;
+use crate::unified_exec::UnifiedExecError;
+use crate::unified_exec::UnifiedExecProcessManager;
+use crate::unified_exec::generate_chunk_id;
+use codex_features::Feature;
+use codex_otel::SessionTelemetry;
+use codex_otel::TOOL_CALL_UNIFIED_EXEC_METRIC;
+use codex_shell_command::is_safe_command::is_known_safe_command;
+use codex_tools::ToolName;
+use codex_utils_output_truncation::approx_token_count;
+
+use super::ExecCommandArgs;
+use super::ExecCommandEnvironmentArgs;
+use super::effective_max_output_tokens;
+use super::get_command;
+use super::post_unified_exec_tool_use_payload;
+
+pub struct ExecCommandHandler;
+
+impl ToolHandler for ExecCommandHandler {
+    type Output = ExecCommandToolOutput;
+
+    fn tool_name(&self) -> ToolName {
+        ToolName::plain("exec_command")
+    }
+
+    fn kind(&self) -> ToolKind {
+        ToolKind::Function
+    }
+
+    fn matches_kind(&self, payload: &ToolPayload) -> bool {
+        matches!(payload, ToolPayload::Function { .. })
+    }
+
+    async fn is_mutating(&self, invocation: &ToolInvocation) -> bool {
+        let ToolPayload::Function { arguments } = &invocation.payload else {
+            tracing::error!(
+                "This should never happen, invocation payload is wrong: {:?}",
+                invocation.payload
+            );
+            return true;
+        };
+
+        let Ok(params) = parse_arguments::<ExecCommandArgs>(arguments) else {
+            return true;
+        };
+        let command = match get_command(
+            &params,
+            invocation.session.user_shell(),
+            &invocation.turn.tools_config.unified_exec_shell_mode,
+            invocation.turn.tools_config.allow_login_shell,
+        ) {
+            Ok(command) => command,
+            Err(_) => return true,
+        };
+        !is_known_safe_command(&command)
+    }
+
+    fn pre_tool_use_payload(&self, invocation: &ToolInvocation) -> Option<PreToolUsePayload> {
+        let ToolPayload::Function { arguments } = &invocation.payload else {
+            return None;
+        };
+
+        parse_arguments::<ExecCommandArgs>(arguments)
+            .ok()
+            .map(|args| PreToolUsePayload {
+                tool_name: HookToolName::bash(),
+                tool_input: serde_json::json!({ "command": args.cmd }),
+            })
+    }
+
+    fn post_tool_use_payload(
+        &self,
+        invocation: &ToolInvocation,
+        result: &Self::Output,
+    ) -> Option<PostToolUsePayload> {
+        post_unified_exec_tool_use_payload(invocation, result)
+    }
+
+    async fn handle(&self, invocation: ToolInvocation) -> Result<Self::Output, FunctionCallError> {
+        let ToolInvocation {
+            session,
+            turn,
+            tracker,
+            call_id,
+            payload,
+            ..
+        } = invocation;
+
+        let arguments = match payload {
+            ToolPayload::Function { arguments } => arguments,
+            _ => {
+                return Err(FunctionCallError::RespondToModel(
+                    "exec_command handler received unsupported payload".to_string(),
+                ));
+            }
+        };
+
+        let manager: &UnifiedExecProcessManager = &session.services.unified_exec_manager;
+        let context = UnifiedExecContext::new(session.clone(), turn.clone(), call_id.clone());
+        let environment_args: ExecCommandEnvironmentArgs = parse_arguments(&arguments)?;
+        let Some(turn_environment) =
+            resolve_tool_environment(turn.as_ref(), environment_args.environment_id.as_deref())?
+        else {
+            return Err(FunctionCallError::RespondToModel(
+                "unified exec is unavailable in this session".to_string(),
+            ));
+        };
+        let cwd = environment_args
+            .workdir
+            .as_deref()
+            .filter(|workdir| !workdir.is_empty())
+            .map_or_else(
+                || turn_environment.cwd.clone(),
+                |workdir| turn_environment.cwd.join(workdir),
+            );
+        let environment = Arc::clone(&turn_environment.environment);
+        let fs = environment.get_filesystem();
+        let args: ExecCommandArgs = parse_arguments_with_base_path(&arguments, &cwd)?;
+        let hook_command = args.cmd.clone();
+        maybe_emit_implicit_skill_invocation(
+            session.as_ref(),
+            context.turn.as_ref(),
+            &hook_command,
+            &cwd,
+        )
+        .await;
+        let process_id = manager.allocate_process_id().await;
+        let command = get_command(
+            &args,
+            session.user_shell(),
+            &turn.tools_config.unified_exec_shell_mode,
+            turn.tools_config.allow_login_shell,
+        )
+        .map_err(FunctionCallError::RespondToModel)?;
+        let command_for_display = codex_shell_command::parse_command::shlex_join(&command);
+
+        let ExecCommandArgs {
+            tty,
+            yield_time_ms,
+            max_output_tokens,
+            sandbox_permissions,
+            additional_permissions,
+            justification,
+            prefix_rule,
+            ..
+        } = args;
+        let max_output_tokens =
+            effective_max_output_tokens(max_output_tokens, turn.truncation_policy);
+
+        let exec_permission_approvals_enabled =
+            session.features().enabled(Feature::ExecPermissionApprovals);
+        let requested_additional_permissions = additional_permissions.clone();
+        let effective_additional_permissions = apply_granted_turn_permissions(
+            context.session.as_ref(),
+            cwd.as_path(),
+            sandbox_permissions,
+            additional_permissions,
+        )
+        .await;
+        let additional_permissions_allowed = exec_permission_approvals_enabled
+            || (session.features().enabled(Feature::RequestPermissionsTool)
+                && effective_additional_permissions.permissions_preapproved);
+
+        // Sticky turn permissions have already been approved, so they should
+        // continue through the normal exec approval flow for the command.
+        if effective_additional_permissions
+            .sandbox_permissions
+            .requests_sandbox_override()
+            && !effective_additional_permissions.permissions_preapproved
+            && !matches!(
+                context.turn.approval_policy.value(),
+                codex_protocol::protocol::AskForApproval::OnRequest
+            )
+        {
+            let approval_policy = context.turn.approval_policy.value();
+            manager.release_process_id(process_id).await;
+            return Err(FunctionCallError::RespondToModel(format!(
+                "approval policy is {approval_policy:?}; reject command — you cannot ask for escalated permissions if the approval policy is {approval_policy:?}"
+            )));
+        }
+
+        let normalized_additional_permissions = match implicit_granted_permissions(
+            sandbox_permissions,
+            requested_additional_permissions.as_ref(),
+            &effective_additional_permissions,
+        )
+        .map_or_else(
+            || {
+                normalize_and_validate_additional_permissions(
+                    additional_permissions_allowed,
+                    context.turn.approval_policy.value(),
+                    effective_additional_permissions.sandbox_permissions,
+                    effective_additional_permissions.additional_permissions,
+                    effective_additional_permissions.permissions_preapproved,
+                    &cwd,
+                )
+            },
+            |permissions| Ok(Some(permissions)),
+        ) {
+            Ok(normalized) => normalized,
+            Err(err) => {
+                manager.release_process_id(process_id).await;
+                return Err(FunctionCallError::RespondToModel(err));
+            }
+        };
+
+        if let Some(output) = intercept_apply_patch(
+            &command,
+            &cwd,
+            fs.as_ref(),
+            context.session.clone(),
+            context.turn.clone(),
+            Some(&tracker),
+            &context.call_id,
+            "exec_command",
+        )
+        .await?
+        {
+            manager.release_process_id(process_id).await;
+            return Ok(ExecCommandToolOutput {
+                event_call_id: String::new(),
+                chunk_id: String::new(),
+                wall_time: std::time::Duration::ZERO,
+                raw_output: output.into_text().into_bytes(),
+                max_output_tokens: Some(max_output_tokens),
+                process_id: None,
+                exit_code: None,
+                original_token_count: None,
+                hook_command: None,
+            });
+        }
+
+        emit_unified_exec_tty_metric(&turn.session_telemetry, tty);
+        match manager
+            .exec_command(
+                ExecCommandRequest {
+                    command,
+                    hook_command: hook_command.clone(),
+                    process_id,
+                    yield_time_ms,
+                    max_output_tokens: Some(max_output_tokens),
+                    cwd,
+                    environment,
+                    network: context.turn.network.clone(),
+                    tty,
+                    sandbox_permissions: effective_additional_permissions.sandbox_permissions,
+                    additional_permissions: normalized_additional_permissions,
+                    additional_permissions_preapproved: effective_additional_permissions
+                        .permissions_preapproved,
+                    justification,
+                    prefix_rule,
+                },
+                &context,
+            )
+            .await
+        {
+            Ok(response) => Ok(response),
+            Err(UnifiedExecError::SandboxDenied { output, .. }) => {
+                let output_text = output.aggregated_output.text;
+                let original_token_count = approx_token_count(&output_text);
+                Ok(ExecCommandToolOutput {
+                    event_call_id: context.call_id.clone(),
+                    chunk_id: generate_chunk_id(),
+                    wall_time: output.duration,
+                    raw_output: output_text.into_bytes(),
+                    max_output_tokens: Some(max_output_tokens),
+                    // Sandbox denial is terminal, so there is no live
+                    // process for write_stdin to resume.
+                    process_id: None,
+                    exit_code: Some(output.exit_code),
+                    original_token_count: Some(original_token_count),
+                    hook_command: Some(hook_command),
+                })
+            }
+            Err(err) => Err(FunctionCallError::RespondToModel(format!(
+                "exec_command failed for `{command_for_display}`: {err:?}"
+            ))),
+        }
+    }
+}
+
+fn emit_unified_exec_tty_metric(session_telemetry: &SessionTelemetry, tty: bool) {
+    session_telemetry.counter(
+        TOOL_CALL_UNIFIED_EXEC_METRIC,
+        /*inc*/ 1,
+        &[("tty", if tty { "true" } else { "false" })],
+    );
+}
